refactor: move seed, handler, and watcher functions out of singleton.go
Reviewer feedback: keep singleton.go minimal by moving
seedRedHatDefaultSignatureIntegration, keyBundleHandler, and
startKeyBundleWatcher into key_bundle.go alongside the types they
operate on. Their tests move from singleton_test.go to
key_bundle_test.go accordingly.

Partially generated by AI.
